refactor: move all model doc types to a dbmodel type file
